pyproject.toml
[project]
name = "mcp-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.135.1",
    "langchain-core>=1.2.17",
    "openai>=2.26.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "sqlparse>=0.5.5",
    "sqlvalidator>=0.0.20",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "uvicorn>=0.41.0",
]
